Changed WPF pattern so Prompt popup appears in front of app
