+++ src/WPF/ArcGISRuntime.WPF.Viewer/MainWindow.xaml.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2018 Esri.
+﻿// Copyright 2021 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
@@ -8,6 +8,7 @@
 // language governing permissions and limitations under the License.
 
 using ArcGISRuntime.Samples.Managers;
+using ArcGISRuntime.Samples.Shared.Managers;
 using ArcGISRuntime.Samples.Shared.Models;
 using Esri.ArcGISRuntime.Security;
 using System;
@@ -23,6 +24,7 @@ namespace ArcGISRuntime.Samples.Desktop
     public partial class MainWindow
     {
         private bool _waitFlag;
+
         public MainWindow()
         {
             InitializeComponent();
@@ -41,11 +43,35 @@ namespace ArcGISRuntime.Samples.Desktop
 
                 // Select the first item
                 samples.First().IsSelected = true;
+
+                Loaded += CheckApiKey;
             }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.ToString(), "Initialization exception occurred.");
             }
+        }
+
+        private async void CheckApiKey(object sender, RoutedEventArgs e)
+        {
+            Loaded -= CheckApiKey;
+
+            // Attempt to load a locally stored API key.
+            await ApiKeyManager.TrySetLocalKey();
+
+            // Check that the current API key is valid.
+            ApiKeyStatus status = await ApiKeyManager.CheckKeyValidity();
+            if (status != ApiKeyStatus.Valid)
+            {
+                PromptForKey();
+            }
+        }
+
+        private void PromptForKey()
+        {
+            var keyPrompt = new Window() { Width = 500, Height = 220, Title = "Edit API key" };
+            keyPrompt.Content = new ApiKeyPrompt();
+            keyPrompt.Show();
         }
 
         private async void categoriesList_SelectionChanged(object sender, SelectionChangedEventArgs e)
@@ -66,7 +92,7 @@ namespace ArcGISRuntime.Samples.Desktop
         private async void categories_SelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
         {
             var context = e.NewValue as TreeViewItem;
-            if (context == null) {return;}
+            if (context == null) { return; }
             var sample = context.DataContext as SampleInfo;
             var category = context.DataContext as SearchableTreeNode;
             if (category != null)
@@ -178,7 +204,7 @@ namespace ArcGISRuntime.Samples.Desktop
             SampleContainer.Visibility = Visibility.Visible;
             DescriptionContainer.Visibility = Visibility.Collapsed;
             CategoriesRegion.Visibility = Visibility.Collapsed;
-            SourceCodeContainer.Visibility = Visibility.Collapsed; 
+            SourceCodeContainer.Visibility = Visibility.Collapsed;
         }
 
         private void ShowDescriptionTab()
